+++ Source/API/Section/PDFSectionColumn+Objects.swift
@@ -288,4 +288,20 @@ public extension PDFSectionColumn {
     func setAbsoluteOffset(_ container: PDFSectionColumnContainer = PDFSectionColumnContainer.left, offset: CGFloat) {
         set(container, absoluteOffset: offset)
     }
+
+    // MARK: - Groups
+
+    /**
+    Adds a group object to the section column in the defined container
+    */
+    func add(_ container: PDFSectionColumnContainer = PDFSectionColumnContainer.left, group: PDFGroup) {
+        objects += [(container, PDFGroupObject(objects: group.objects,
+                                               allowsBreaks: group.allowsBreaks,
+                                               isFullPage: false,
+                                               backgroundColor: group.backgroundColor,
+                                               backgroundImage: group.backgroundImage,
+                                               backgroundShape: group.backgroundShape,
+                                               outline: group.outline,
+                                               padding: group.padding))]
+    }
 }